Abstracted SQL data from the server file
